pyproject.toml
[project]
name = "langchain-llm-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "langchain-core (>=0.3.29,<0.4.0)",
    "fastapi[standard] (>=0.115.6,<0.116.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
